+++ src/container/SequentialContainer/Base/RandomIterator.ts
@@ -1,8 +1,9 @@
 import { ContainerIterator, IteratorType } from '@/container/ContainerBase';
-import $checkWithinAccessParams from '@/utils/checkParams.macro';
+import SequentialContainer from '@/container/SequentialContainer/Base/index';
 import { throwIteratorAccessError } from '@/utils/throwError';
 
 export abstract class RandomIterator<T> extends ContainerIterator<T> {
+  abstract readonly container: SequentialContainer<T>;
   /**
    * @internal
    */
@@ -10,30 +11,12 @@ export abstract class RandomIterator<T> extends ContainerIterator<T> {
   /**
    * @internal
    */
-  protected readonly _size: () => number;
-  /**
-   * @internal
-   */
-  protected readonly _getElementByPos: (pos: number) => T;
-  /**
-   * @internal
-   */
-  protected readonly _setElementByPos: (pos: number, element: T) => void;
-  /**
-   * @internal
-   */
-  constructor(
+  protected constructor(
     index: number,
-    size: () => number,
-    getElementByPos: (pos: number) => T,
-    setElementByPos: (pos: number, element: T) => void,
     iteratorType?: IteratorType
   ) {
     super(iteratorType);
     this._node = index;
-    this._size = size;
-    this._getElementByPos = getElementByPos;
-    this._setElementByPos = setElementByPos;
     if (this.iteratorType === IteratorType.NORMAL) {
       this.pre = function () {
         if (this._node === 0) {
@@ -43,7 +26,7 @@ export abstract class RandomIterator<T> extends ContainerIterator<T> {
         return this;
       };
       this.next = function () {
-        if (this._node === this._size()) {
+        if (this._node === this.container.size()) {
           throwIteratorAccessError();
         }
         this._node += 1;
@@ -51,7 +34,7 @@ export abstract class RandomIterator<T> extends ContainerIterator<T> {
       };
     } else {
       this.pre = function () {
-        if (this._node === this._size() - 1) {
+        if (this._node === this.container.size() - 1) {
           throwIteratorAccessError();
         }
         this._node += 1;
@@ -67,12 +50,10 @@ export abstract class RandomIterator<T> extends ContainerIterator<T> {
     }
   }
   get pointer() {
-    $checkWithinAccessParams!(this._node, 0, this._size() - 1);
-    return this._getElementByPos(this._node);
+    return this.container.getElementByPos(this._node);
   }
   set pointer(newValue: T) {
-    $checkWithinAccessParams!(this._node, 0, this._size() - 1);
-    this._setElementByPos(this._node, newValue);
+    this.container.setElementByPos(this._node, newValue);
   }
   // @ts-ignore
   pre(): this;