+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog and this project adheres to Semantic Versioning.
 
+## [4.3.0] - 2023.01.20
+
+### Added
+
+- Add public member `container` to `Iterator` which means the container that the iterator pointed to.
+
+### Changed
+
+- Reimplement `Queue`, separate `Queue` from `Deque`.
+
 ## [4.2.0] - 2022.11.20
 
 ### Changed